#41 determine complete config diff
